+++ TowerDefense/Tile.h
@@ -8,7 +8,7 @@
 class Tile
 {
 private:
-	sf::Vector2f position;//左上角点的位置
+	sf::Vector2i position;//左上角点的位置
 	int width;
 	int height;
 	int cooldown;//剩余的冷却时间，默认为正常状态，值为0.若被污染，倒计时TILE_COOLDOWN_TIME秒恢复
